Fixed plist Purging; corrected all warnings in 32 and 64 bit versions. Shogi: Computer nearly beats me now.

+++ BG/Core/BGCorePosition.cpp
@@ -9,7 +9,7 @@ namespace BoardGamesCore
     std::size_t z{};
     for (auto& p : pieces)
     {
-      z ^= p + 0x9e3779b9 + (z << 6) + (z >> 2);
+      z ^= static_cast<size_t>(p) + 0x9e3779b9 + (z << 6) + (z >> 2);
 //      z ^= pMap->GetPiece(p)->GetHash() + 0x9e3779b9 + (z << 6) + (z >> 2);
     }
     return hash = z;
